distribution implementations added

+++ java/src/contention/abstractions/Distribution.java
@@ -1,5 +1,5 @@
 package contention.abstractions;
 
 public interface Distribution {
-    long next();
+    int next();
 }
